implement image-processing micro-bench

+++ micro-bench/src/sequential.rs
@@ -1,39 +1,39 @@
-use std::time::SystemTime;
-use std::io::Write;
 use std::fs::File;
+use std::io::Write;
+use std::time::SystemTime;
 
 pub fn sequential(size: usize, iter_size1: i32, iter_size2: i32) {
-
     let start = SystemTime::now();
 
-    let init_a = -2.125 as f64;
-    let init_b = -1.5 as f64;
-    let range = 3.0 as f64;
+    let init_a = -2.125;
+    let init_b = -1.5;
+    let range = 3.0;
     let step = range / (size as f64);
 
-    let mut m = vec![0; size*size];
+    let mut m = vec![0; size * size];
     let iter = iter_size1 + iter_size2;
 
-    for i in 0 .. size{
+    for i in 0..size {
         let im = init_b + (step * (i as f64));
 
-        for j in 0 .. size {
+        for j in 0..size {
             let mut a = init_a + step * j as f64;
             let cr = a;
             let mut b = im;
             let mut k = 0;
 
-            for ii in 0 .. iter {
+            for ii in 0..iter {
                 let a2 = a * a;
                 let b2 = b * b;
-                if (a2 + b2) > 4.0 {break;}
+                if (a2 + b2) > 4.0 {
+                    break;
+                }
                 b = 2.0 * a * b + im;
                 a = a2 - b2 + cr;
                 k = ii;
             }
-            
-            m[i*size+j] = (255 as f64 - ((k as f64 * 255 as f64 / (iter as f64)))) as u8;
-        
+
+            m[i * size + j] = (255.0 - (k as f64 * 255.0 / (iter as f64))) as u8;
         }
     }
 
@@ -44,3 +44,4 @@ pub fn sequential(size: usize, iter_size1: i32, iter_size2: i32) {
     let mut buffer = File::create("result_sequential.txt").unwrap();
     buffer.write_all(&m).unwrap();
 }
+
